+++ uni_modules/uni-cmark/native/md2html.h
@@ -18,6 +18,10 @@ UNI_CMARK_INTERNAL char *uni_cmark_markdown_to_html(
     const uint8_t *markdown, size_t markdown_length, size_t *html_length);
 UNI_CMARK_INTERNAL void uni_cmark_free_html(char *html);
 
+UNI_CMARK_INTERNAL char *uni_cmark_markdown_to_json(
+    const uint8_t *markdown, size_t markdown_length, size_t *json_length);
+UNI_CMARK_INTERNAL void uni_cmark_free_json(char *json);
+
 #ifdef __cplusplus
 }
 #endif